Fix property search pages ss template location. Add editor config file.
